Pull in patches from FFyte

+++ engines/scumm/scumm_v0.h
@@ -81,7 +81,7 @@ protected:
 
 	void setupScummVars() override;
 	void resetScummVars() override;
-	void scummLoop(int delta) override;
+	void scummLoop(byte delta) override;
 	void decodeParseString() override;
 
 	void processInput() override;
